### 2025-04-22 #8041 @skanwal
Hi Chris. Yes, I can.
