### 2025-04-22 2232c9b
Update license to CRAN standards
