### 2025-04-22 1a7dbd8
Remove non-standard top level directory JOSS
