### 2025-04-22 5f3e958
Fix rstantools CRAN check error
