+++ R/swolca_var_adjust.R
@@ -179,7 +179,6 @@ grad_par <- function(pwts, svydata, stan_mod, stan_data, par_stan, u_pars) {
 #' @importFrom rstan sampling unconstrain_pars grad_log_prob constrain_pars
 #' @importFrom survey svydesign as.svrepdesign withReplicates
 #' @importFrom Matrix nearPD
-#' @importFrom rstantools rstan_config
 #' @export
 #' 
 #' @references 
