+++ DESCRIPTION
@@ -3,15 +3,20 @@ Type: Package
 Title: Bayesian Survey Clustering
 Version: 1.0
 Authors@R: c(
-    person(given = "Stephanie", family = "Wu", email = "stephwu18@gmail.com", role = c("aut", "cre"), comment = c(ORCID = "0000-0001-5110-8407")), 
-    person(given = "Matthew", family = "Williams", email = "mrwilliams@rti.org", role = "ctb", comment = c(ORCID = "0000-0001-8894-1240")),
-    person(given = "Terrance", family = "Savitsky", email = "savitsky.terrance@bls.gov", role = "ctb", comment = c(ORCID = "0000-0003-1843-3106")), 
-    person(given = "Briana", family = "Stephenson", email = "bstephenson@hsph.harvard.edu", role = "ctb", comment = c(ORCID = "0000-0002-6147-1039")))
+    person("Stephanie", "Wu", , "isytchev@hsph.harvard.edu", role = c("aut", "cre"),
+           comment = c(ORCID = "0000-0001-5110-8407")),
+    person("Matthew", "Williams", , "mrwilliams@rti.org", role = "ctb",
+           comment = c(ORCID = "0000-0001-8894-1240")),
+    person("Terrance", "Savitsky", , "savitsky.terrance@bls.gov", role = "ctb",
+           comment = c(ORCID = "0000-0003-1843-3106")),
+    person("Briana", "Stephenson", , "bstephenson@hsph.harvard.edu", role = "ctb",
+           comment = c(ORCID = "0000-0002-6147-1039"))
+  )
 Description: An R package for running Bayesian clustering methods using survey data. Detailed instructions can be found in the vignette. A Bayesian latent class analysis is available for eliciting underlying pattern profiles incorporating survey sampling weights and other survey design elements. Options also exist for relating the pattern profiles to a binary outcome, either through a two-step approach after cluster is completed, or in a one-step approach where creation of the pattern profiles is directly informed by the outcome. Summary and plotting functions for visualizing output are available, as are diagnostic functions for examining convergence of the sampler. The package includes an example dataset from the National Health and Nutrition Examination Survey with dietary intake and hypertension data for low-income women in the United States, as well as a simulated dataset to help with familiarization of the package.
 License: MIT + file LICENSE
 Encoding: UTF-8
 LazyData: true
-Imports: 
+Imports:
     dplyr,
     e1071,
     ggplot2,
@@ -27,7 +32,6 @@ Imports:
     RcppTN,
     rlang,
     rstan (>= 2.21.8),
-    rstantools (>= 2.3.1.1),
     stats,
     survey (>= 4.2.1),
     tidyr,
@@ -52,6 +56,7 @@ Suggests:
     testthat (>= 3.0.0),
     knitr,
     rmarkdown,
+    rstantools (>= 2.3.1.1),
     devtools,
     stringr,
     SimCorMultRes
