### 2025-06-09 #8382 @editorialbot
[REVIEW]: baysc: An R package for Bayesian survey clustering
**Submitting author:** @smwu (<a href="http://orcid.org/0000-0001-5110-8407">Stephanie Wu</a>)
**Repository:** https://github.com/smwu/baysc
**Branch with paper.md** (empty if default branch):
**Version:** v1.0.0
