+++ JOSS/baysc.bib
@@ -1,0 +1,224 @@
+@article{stephenson2020empirically,
+  title={Empirically Derived Dietary Patterns Using Robust Profile Clustering in the Hispanic Community Health Study/Study of Latinos},
+  author={Stephenson, Briana JK and Sotres-Alvarez, Daniela and Siega-Riz, Anna-Maria and Mossavar-Rahmani, Yasmin and Daviglus, Martha L and Van Horn, Linda and Herring, Amy H and Cai, Jianwen},
+  journal={The Journal of Nutrition},
+  volume={150},
+  number={10},
+  pages={2825--2834},
+  year={2020},
+  publisher={Oxford University Press}
+}
+
+@article{lanza2016latent,
+  title={Latent class analysis for developmental research},
+  author={Lanza, Stephanie T and Cooper, Brittany R},
+  journal={Child Development Perspectives},
+  volume={10},
+  number={1},
+  pages={59--64},
+  year={2016},
+  publisher={Wiley Online Library}
+}
+
+@misc{nchs2023homepage,
+    title={National Health and Nutrition Examination Survey Home Page},
+    author={{National Center for Health Statistics}},
+    year={2023},
+    publisher={Centers for Disease Control and Prevention, National Center for Health Statistics, Atlanta, GA},
+    url={https://www.cdc.gov/nchs/nhanes.htm},
+    urldat={2023-07-23}
+}
+
+@article{wu2024derivation,
+  title={Derivation of outcome-dependent dietary patterns for low-income women obtained from survey data using a supervised weighted overfitted latent class analysis},
+  author={Wu, Stephanie M and Williams, Matthew R and Savitsky, Terrance D and Stephenson, Briana JK},
+  journal={Biometrics},
+  volume={80},
+  number={4},
+  pages={ujae122},
+  year={2024},
+  publisher={Oxford University Press}
+}
+
+@article{stephenson2023identifying,
+    author = {Stephenson, Briana J K and Wu, Stephanie M and Dominici, Francesca},
+    title = "{Identifying dietary consumption patterns from survey data: a Bayesian nonparametric latent class model}",
+    journal = {Journal of the Royal Statistical Society Series A: Statistics in Society},
+    pages = {qnad135},
+    year = {2023},
+    month = {12},
+    abstract = "{Dietary assessments provide the snapshots of population-based dietary habits. Questions remain about how generalisable those snapshots are in national survey data, where certain subgroups are sampled disproportionately. We propose a Bayesian overfitted latent class model to derive dietary patterns, accounting for survey design and sampling variability. Compared to standard approaches, our model showed improved identifiability of the true population pattern and prevalence in simulation. We focus application of this model to identify the intake patterns of adults living at or below the 130\\% poverty income level. Five dietary patterns were identified and characterised by reproducible code/data made available to encourage further research.}",
+    issn = {0964-1998},
+    doi = {10.1093/jrsssa/qnad135},
+    url = {https://doi.org/10.1093/jrsssa/qnad135},
+    eprint = {https://academic.oup.com/jrsssa/advance-article-pdf/doi/10.1093/jrsssa/qnad135/54350625/qnad135.pdf},
+}
+
+@article{albert1993bayesian,
+  title={Bayesian analysis of binary and polychotomous response data},
+  author={Albert, James H and Chib, Siddhartha},
+  journal={Journal of the American Statistical Association},
+  volume={88},
+  number={422},
+  pages={669--679},
+  year={1993},
+  publisher={Taylor \& Francis}
+}
+
[…]
