### 2025-07-12 #8382 @smwu
@skanwal, my apologies for the delay and oversight. The paper directory had been removed to comply with CRAN submission. I have created a separate branch named 'joss-paper' that has the paper included. Would it be possible to proceed with the review in this branch? [https://github.com/smwu/baysc/tree/joss-paper](https://github.com/smwu/baysc/tree/joss-paper)
