### 2025-07-23 smwu/baysc#2 @CalebAbhulimhen
LICENSE doesn't meet JOSS standards.
According to the JOSS criteria, the LICENSE should be a plain-text file.

In your repo, you have a LICENSE and LICENSE.md.

- The content of LICENSE is not an [OSI-approved license](https://opensource.org/license)
- Although the content of LICENSE.md is a MIT license, the file name shouldn't include the ".md" extension.
