### 2025-07-27 72c68f3
Updated license to align with JOSS requirements
