### 2025-08-16 smwu/baysc#2 @CalebAbhulimhen
Hi, the plain text [LICENSE](https://github.com/smwu/baysc/blob/main/LICENSE) file still doesn't contain the MIT license. Please let me know if you're referring to another file in your comment.
