### 2025-09-02 smwu/baysc#3 @CalebAbhulimhen
JOSS Review - Update the Installing Instructions.
Hi,

Can you update the installation instructions in the README to include [this section](https://github.com/smwu/baysc/blob/7c62364f9a2af2f66e0307b0c43d772070ef73de/vignettes/baysc.Rmd#L88-L89) of the vignette?

I ran into the first issue, and I believe @larryshamalama ran into the second one.

Thanks!
