### 2025-09-02 #8382 @CalebAbhulimhen
Hi @smwu, great work on the baysc package!

Overall, it looks good to me. I created some issues on the repo to track the outstanding checklist items:

- License: https://github.com/smwu/baysc/issues/2
- Installation: https://github.com/smwu/baysc/issues/3

Also, +1 to @larryshamalama's comment on expanding the software paper to include an "How to" section.

I'll approve the review once these issues are resolved.

Thanks!
