+++ CITATION.cff
@@ -11,7 +11,6 @@ authors:
   - given-names: Stephanie
     family-names: Wu
     email: stephanie.wu@ucl.ac.uk
-    name-particle: Stephanie
     affiliation: UCL
     orcid: 'https://orcid.org/0000-0001-5110-8407'
   - given-names: Matthew
