### 2025-09-10 65f9ca9
Updated citation and removed github .cff file
