### 2025-09-10 591bee0
Updated package version and Makevars files so RcppArmadillo uses current Armadillo
