+++ DESCRIPTION
@@ -1,7 +1,7 @@
 Type: Package
 Package: baysc
-Title: Bayesian Survey Clustering
-Version: 1.0
+Title: BAYesian Survey Clustering
+Version: 0.1.0
 Authors@R: c(
     person("Stephanie", "Wu", , "stephanie.wu@ucl.ac.uk", role = c("aut", "cre"),
            comment = c(ORCID = "0000-0001-5110-8407")),
