### 2025-09-15 24b0c3c
Updated JOSS branch citation, description, and RcppArmadillo compiler flag
