### 2025-09-16 61db564
Bug fix for convert_mix_to_ref for wolca_svyglm summary
