### 2025-09-24 6711f12
Updated the JOSS paper with usage examples
