### 2025-09-24 7cd942e
Minor edits to vignette to fix typos
