+++ vignettes/baysc.Rmd
@@ -138,7 +138,7 @@ glm_form <- "~ age_cat + racethnic + smoker + physactive"
 The following sections describe how to use the main functions in the `baysc` package for running survey-weighted Bayesian LCA. If no outcome information is necessary when clustering, use the unsupervised weighted OLCA, implemented in the `wolca()` function ([Section 2.2.1](#wolca)). If clustering is to be informed by a binary outcome so that latent classes are found that specifically drive the outcome of interest, use the supervised weighted OLCA, implemented in the `swolca()` function ([Section 2.2.3](#swolca)). 
 
 ### Use `wolca()` and `wolca_var_adjust()` to derive dietary exposure clusters {#wolca}
-We use the `wolca()` function to run an unsupervised weighted overfitted latent class analysis and obtain pattern profiles. The `wolca_var_adjust()` function is used after to apply a post-processing adjustment that allows for correct estimation of the variability of the parameter estimates. If this adjustment is not applied, the estimated variances will be too small. We run the sampler for 100 iterations as an example, with a total runtime of around 5 minutes for `wolca()` and `wolca_var_adjust()`. Typically, at least 20000 iterations should be used for `wolca()`, corresponding to a runtime of around 2 hours.
+We use the `wolca()` function to run an unsupervised weighted overfitted latent class analysis and obtain pattern profiles. The `wolca_var_adjust()` function is used after to apply a post-processing adjustment that allows for correct estimation of the variability of the parameter estimates. If this adjustment is not applied, the estimated variances will be too small. We run the sampler for 300 iterations as an example, with a total runtime of around 5 minutes for `wolca()` and `wolca_var_adjust()`. Typically, at least 20000 iterations should be used for `wolca()`, corresponding to a runtime of around 2 hours.
 
 When running the model, there are two samplers that can be run: an adaptive sampler and a fixed sampler. The adaptive sampler is used to determine the number of latent classes using a data-driven process with a sparsity-inducing Dirichlet prior. Once the number of latent classes is determined, the fixed sampler is run using the set number of classes and produces output for estimation and inference. Note that the final number of classes determined by the fixed sampler may be smaller than the number of classes obtained from the adaptive sampler. 
 
@@ -162,7 +162,7 @@ Parameters used for `wolca()`:
 - `stratum_id`: nx1 vector of stratum indicators.
 - `run_sampler = "both"`: Run both the adaptive and fixed sampler (default). Other options are `"adapt"` for only the adaptive sampler, or `"fixed"` for only the fixed sampler when the number of latent classes is known a priori.
 - `K_max = 30`: The maximum number of latent classes is set to 30 (default). Typically, the true number will be much smaller.
-- `adapt_seed = 888`: Random seed for adaptive sampler, set to 888 for this example but with default `NULL`. 
+- `adapt_seed = 999`: Random seed for adaptive sampler, set to 999 for this example but with default `NULL`. 
 - `class_cutoff = 0.05`: The minimum size of each latent class, with default set to 5\% of the population.
 - `n_runs = 300`: Number of MCMC sampler iterations to run. Set to 300 for this example but with default 20000. *Typically, at least 20000 iterations should be used to allow the sampler to converge.* 
 - `burn = 150`: Number of iterations for burn-in. Set to 150 for this example but with default 10000. 
@@ -173,13 +173,13 @@ Parameters used for `wolca()`:
 Parameters used for `wolca_var_adjust()`: 
 
 - `res = res_wolca_unadj`: Results from calling `wolca()`.
-- `adjust_seed = 888`: Random seed for the adjustment procedure, set to 888 for this example but with default `NULL`.
[…]
