### 2025-09-24 1c5bf43
Added supplementary installation instructions to README and updated citation
