+++ README.md
@@ -6,20 +6,19 @@ An R package for running Bayesian supervised and unsupervised clustering methods
 
 **Contributors**: Matthew R. Williams ([mrwilliams\@rti.org](mailto:mrwilliams@rti.org){.email}); Terrance D. Savitsky ([savitsky.terrance\@bls.gov](mailto:savitsky.terrance@bls.gov){.email}); Briana J.K. Stephenson ([bstephenson\@hsph.harvard.edu](mailto:bstephenson@hsph.harvard.edu){.email})
 
-**Citation**
-Wu, S. M., Williams, M. R., Savitsky, T. D., & Stephenson, B. J. (2024). Derivation of outcome-dependent dietary patterns for low-income women obtained from survey data using a supervised weighted overfitted latent class analysis. Biometrics, 80(4), ujae122, <https://doi.org/10.1093/biomtc/ujae122>
+**Citation**: Wu S, Williams M, Savitsky T, Stephenson B (2025). *baysc: BAYesian Survey Clustering*. R package version 0.1.0, <https://github.com/smwu/baysc>.
 
-## Table of content
+## Table of contents
 
 -   [1. Installation](#id-section1)
 -   [2. Overview](#id-section2)
 -   [3. Functions](#id-section3)
--   [2. Data](#id-section4)
--   [2. Example](#id-section5)
--   [2. Help](#id-section6)
+-   [4. Data](#id-section4)
+-   [5. Example](#id-section5)
+-   [6. Supplementary](#id-section6)
+-   [7. Help](#id-section7)
 
 <div id='id-section1'/>
-
 ## Installation
 
 ``` r
@@ -31,20 +30,26 @@ devtools::install_github("smwu/baysc")
 library(baysc)
 ```
 
-<div id='id-section2'/>
+During installation, the following errors may arise:
 
+-   *No package called 'rstantools'*: Please install the `rstantools` package using `install.packages("rstantools")`.
+-   *Library 'gfortran' not found*: This is a compiler configuration issue that can arise when using Rcpp on Mac computers with Apple silicon (e.g., M1-M4 chips). Users may need to install Xcode, GNU Fortran, and OpenMP, and edit the `~/.R/Makevars` file. For more details, see the "Supplementary" section below.
+-   *Library 'emutls_w' not found*: This is a toolchain mismatch issue that can arise when using `Rcpp`-dependent packages on Mac computers with Apple silicon (e.g., M1-M4 chips). Users may need to install gfortran and edit the `~/.R/Makevars` file. For more details, see the "Supplementary" section below.
+</div>
+
+<div id='id-section2'/>
 ## Overview
 
 `baysc` is an R package for running Bayesian clustering methods on survey data. A Bayesian latent class analysis (LCA), termed the Weighted Overfitted Latent Class Analysis (WOLCA), is available for eliciting underlying cluster patterns from multivariate categorical data, incorporating survey sampling weights and other survey design elements. Options also exist for relating the patterns to a binary outcome, either by using a two-step approach that applies WOLCA and then runs a survey-weighted regression, or by utilizing a one-step supervised approach where creation of the cluster patterns is directly informed by the outcome, referred to as the Supervised Weighted Overfitted Latent Class Analysis (SWOLCA). More information about the models can be found in the paper linked above. Summary and plotting functions for visualizing output are also available, as are diagnostic functions for examining convergence of the sampler.
+</div>
 
 <div id='id-section3'/>
-
 ## Functions
 
[…]
