+++ README.md
@@ -2,9 +2,9 @@
 
 An R package for running Bayesian supervised and unsupervised clustering methods on survey data.
 
-**Maintainer**: Stephanie M. Wu ([swu\@g.harvard.edu](mailto:swu@g.harvard.edu){.email})
+**Maintainer**: Stephanie M. Wu ([stephanie.wu@ucl.ac.uk](mailto:stephanie.wu@ucl.ac.uk))
 
-**Contributors**: Matthew R. Williams ([mrwilliams\@rti.org](mailto:mrwilliams@rti.org){.email}); Terrance D. Savitsky ([savitsky.terrance\@bls.gov](mailto:savitsky.terrance@bls.gov){.email}); Briana J.K. Stephenson ([bstephenson\@hsph.harvard.edu](mailto:bstephenson@hsph.harvard.edu){.email})
+**Contributors**: Matthew R. Williams ([mrwilliams\@rti.org](mailto:mrwilliams@rti.org)); Terrance D. Savitsky ([savitsky.terrance\@bls.gov](mailto:savitsky.terrance@bls.gov)); Briana J.K. Stephenson ([bstephenson\@hsph.harvard.edu](mailto:bstephenson@hsph.harvard.edu))
 
 **Citation**: Wu S, Williams M, Savitsky T, Stephenson B (2025). *baysc: BAYesian Survey Clustering*. R package version 0.1.0, <https://github.com/smwu/baysc>.
 
@@ -19,6 +19,7 @@ An R package for running Bayesian supervised and unsupervised clustering methods
 -   [7. Help](#id-section7)
 
 <div id='id-section1'/>
+  
 ## Installation
 
 ``` r
@@ -33,36 +34,44 @@ library(baysc)
 During installation, the following errors may arise:
 
 -   *No package called 'rstantools'*: Please install the `rstantools` package using `install.packages("rstantools")`.
--   *Library 'gfortran' not found*: This is a compiler configuration issue that can arise when using Rcpp on Mac computers with Apple silicon (e.g., M1-M4 chips). Users may need to install Xcode, GNU Fortran, and OpenMP, and edit the `~/.R/Makevars` file. For more details, see the "Supplementary" section below.
+-   *Library 'gfortran' not found*: This is a compiler configuration issue that can arise when using `Rcpp` on Mac computers with Apple silicon (e.g., M1-M4 chips). Users may need to install Xcode, GNU Fortran, and OpenMP, and edit the `~/.R/Makevars` file. For more details, see the "Supplementary" section below.
 -   *Library 'emutls_w' not found*: This is a toolchain mismatch issue that can arise when using `Rcpp`-dependent packages on Mac computers with Apple silicon (e.g., M1-M4 chips). Users may need to install gfortran and edit the `~/.R/Makevars` file. For more details, see the "Supplementary" section below.
 </div>
 
 <div id='id-section2'/>
+  
 ## Overview
 
-`baysc` is an R package for running Bayesian clustering methods on survey data. A Bayesian latent class analysis (LCA), termed the Weighted Overfitted Latent Class Analysis (WOLCA), is available for eliciting underlying cluster patterns from multivariate categorical data, incorporating survey sampling weights and other survey design elements. Options also exist for relating the patterns to a binary outcome, either by using a two-step approach that applies WOLCA and then runs a survey-weighted regression, or by utilizing a one-step supervised approach where creation of the cluster patterns is directly informed by the outcome, referred to as the Supervised Weighted Overfitted Latent Class Analysis (SWOLCA). More information about the models can be found in the paper linked above. Summary and plotting functions for visualizing output are also available, as are diagnostic functions for examining convergence of the sampler.
[…]
