### 2025-09-24 98b5889
Added automatic GitHub R CMD check and unit tests using usethis package
