### 2025-09-24 b0190a7
Update README.md usethis R CMD check to point to joss-paper branch
