### 2025-09-24 5f794cc
Created pre-compiled vignette and updated .Rbuildignore
