### 2025-09-26 1cd1089
Minor updates to paper headings
