### 2025-09-26 8a7b045
Minor changes to static vignette
