+++ DESCRIPTION
@@ -1,6 +1,6 @@
 Type: Package
 Package: baysc
-Title: BAYesian Survey Clustering
+Title: Bayesian Survey Clustering
 Version: 0.1.0
 Authors@R: c(
     person("Stephanie", "Wu", , "stephanie.wu@ucl.ac.uk", role = c("aut", "cre"),
