### 2025-09-26 2e0a8df
Added missing markdown package for vignette building
