+++ DESCRIPTION
@@ -55,6 +55,7 @@ Suggests:
     knitr,
     rmarkdown,
     markdown,
+    rbibutils,
     rstantools (>= 2.3.1.1),
     SimCorMultRes,
     stringr,
