### 2026-01-21 #8382 @smwu
Hi @skanwal I have updated the release and Zenodo archive with the latest information. The new DOI is https://doi.org/10.5281/zenodo.18325950 and the version tag is v0.1.1. Please let me know if there are any additional actions that I should take. Many thanks!
