### 2026-02-20 #8382 @smwu
@skanwal It seems that the compilation may be having an issue with the Zenodo archive link, which includes a nested link. Perhaps a solution could be to set the archive as https://doi.org/10.5281/zenodo.18234759 instead?
