### 2026-02-20 #8382 @danielskatz
👋 @smwu - note that the compile problem isn't related to the archive - the generate pdf command doesn't involve the archive at all.
