### 2026-02-20 #8382 @danielskatz
👋 @openjournals/dev - Can you take a look at this pdf generation problem?
