### 2026-02-27 #8382 @skanwal
@arfon - noticed an error in the article proof on github https://github.com/openjournals/joss-papers/blob/joss.08382/joss.08382/10.21105.joss.08382.pdf
