### 2026-03-06 #8382 @crvernon
@smwu - Thank you for your work on this submission.  However, right now the word count is way too high for a JOSS publication.  I would advise a simple fix by removing the "Usage" section completely from the paper and putting it into documentation that is accessible from your repository or a GitHub pages site.  Then you can simply note in the paper that usage is fully described in those docs via a link.

Let me know once this has been addressed and I'll check the remainder of the submission.  Thank you.
