### 2026-03-06 #8382 @smwu
@crvernon Thank you for your feedback. Would it be possible to know the maximum or recommend word count for a JOSS publication? The "Usage" section was added based on a reviewer request to extend the length of the paper and include information on data format, model fit, and visualization. The contents of that section are already included in the package vignette in the repository. Would you recommend trying to shorten the section to still satisfy the reviewer comments or simply including a link to the vignette? Thank you very much for your guidance.
