### 2026-03-07 #8382 @crvernon
@smwu - Since this submission qualifies under our pre-2026 requirements (things changed at the turn of the year a bit), the JOSS submission docs state "The paper should be between 250-1000 words. Authors submitting papers significantly longer than 1000 words may be asked to reduce the length of their paper." as found [here](https://joss.readthedocs.io/en/pre-2026-submission/paper.html#what-should-my-paper-contain).

I would prefer simply pointing to a link to the vignette because we want people when reading your paper to go to your repository as open-source projects are meant to keep growing.  If lots of usage examples are in your paper, they are set in stone and don't offer the ability to change as your software does.  I appreciate the reviewers comments on this note and don't want to discount their perspective; however, my recommended format meets the structure of JOSS more closely.
