### 2026-03-11 #8382 @smwu
@crvernon Apologies for the delay, and thanks so much for the word count guidance. I have updated the paper per your recommendations and propagated the updates to the GitHub repo and Zenodo archive. Is there anything else that I should change?
