### 2026-03-11 #8382 @crvernon
@smwu thank you for making the changes to the paper.  It is looking good.  I just need you to address the following before I accept:

In you paper:

- [x] LINE 67:  "r" as in the R programming language should be capitalized.  You can do this in your bib file by placing curly brackets around the characters you wish to maintain the formatting of.
- [x] LINE 71:  "r" should be capitalized.
- [x] LINE 74:  The "d" in "dirichlet" should be capitalized.
- [x] LINE 79:  "r" should be capitalized.
- [x] LINE 82:  "r" should be capitalized.
- [x] LINE 86:  "r" should be capitalized.
- [x] LINE 95:  "r" should be capitalized.

Let me know when these changes have been made, and after I confirm, I'll accept.
