### 2026-03-11 #8382 @smwu
@crvernon I've made the requested capitalization changes to the paper and pushed the latest version to GitHub. Should I create a new GitHub release and Zenodo archive version as well?
